+++ engine/src/communication/CommunicationInterface/messageSender.hpp
@@ -8,7 +8,6 @@
 
 #include "ExceptionHandling/BlazingThread.h"
 #include "node.hpp"
-#include "serializer.hpp"
 #include "execution_graph/logic_controllers/CacheMachine.h"
 #include "utilities/ctpl_stl.h"
 #include "protocols.hpp"
@@ -26,7 +25,6 @@ public:
 	 * @brief Constructs a message_sender
 	 *
 	 * @param output_cache The cache machine from where to obtain the data to send
-	 * @param input_cache The cache machine there the node receives data (not acutally used by this class, but this class maintains its scope)
 	 * @param node_address_map A map from node id to Node
 	 * @param num_threads Number of threads the message_sender will use to send data concurrently
 	 * @param context The ucp_context_h
@@ -35,30 +33,27 @@ public:
 	 * @param protocol The comm::blazing_protocol 
 	 */
 	message_sender(std::shared_ptr<ral::cache::CacheMachine> output_cache,
-		std::shared_ptr<ral::cache::CacheMachine> input_cache,
 		const std::map<std::string, node> & node_address_map,
 		int num_threads,
 		ucp_context_h context,
 		ucp_worker_h origin,
 		int ral_id,
-		comm::blazing_protocol protocol);
+		comm::blazing_protocol protocol,
+		bool require_acknowledge);
 
 	static void initialize_instance(std::shared_ptr<ral::cache::CacheMachine> output_cache,
-		std::shared_ptr<ral::cache::CacheMachine> input_cache,
 		std::map<std::string, node> node_address_map,
 		int num_threads,
 		ucp_context_h context,
 		ucp_worker_h origin_node,
 		int ral_id,
-		comm::blazing_protocol protocol);
+		comm::blazing_protocol protocol,
+    	bool require_acknowledge);
 
 	std::shared_ptr<ral::cache::CacheMachine> get_output_cache(){
 		return output_cache;
 	}
-	std::shared_ptr<ral::cache::CacheMachine> get_input_cache(){
-		return input_cache;
-	}
-
+	
 	/**
 	 * @brief A polling function that listens on a cache for data and send it off via some protocol
 	 */
@@ -68,13 +63,13 @@ private:
 
 	ctpl::thread_pool<BlazingThread> pool;
 	std::shared_ptr<ral::cache::CacheMachine> output_cache;
-	std::shared_ptr<ral::cache::CacheMachine> input_cache;
 	std::map<std::string, node> node_address_map;
 	blazing_protocol protocol;
 	ucp_worker_h origin;
 	size_t request_size;
 	int ral_id;
 	bool polling_started{false};
+	bool require_acknowledge;
 };
 
 }  // namespace comm